layoutからpageファイルへ移行
