+++ src/app/page.tsx
@@ -1,7 +1,27 @@
-export default function Home() {
+'use client'
+
+import Form from '@/components/Form';
+import { useRouter } from 'next/navigation';
+import './globals.css';
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  const router = useRouter();
+
+  const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+    router.push('/start');
+  };
   return (
-    <>
-      <p>hello, world</p>
-    </>
+    <html lang="ja">
+      <body className="flex flex-col items-center justify-center min-h-screen p-4">
+        <h1 className="text-3xl font-bold mb-6 text-blue-500">SNSスカウター</h1>
+        <Form onSubmit={handleSubmit} />
+      </body>
+      {children}
+    </html>
   );
 }
